+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG
 
-## Next
+## Unreleased
 
+- Fix `HighlightedCard` imageFit issues.
 - **Breaking Change**: Remove `Bun` component
 
 ## v5.3.2 - 2020-08-28